update warning & buildsets

+++ libhsync/sync_client/sync_info_client.cpp
@@ -505,7 +505,7 @@ static int _TNewDataSyncInfo_open(TNewDataSyncInfo* self,const hpatch_TStreamInp
             
             check(inStream->read(inStream,curPos,codeBuf,codeBuf+compressDataSize),kSyncClient_newSyncInfoDataError);
             decompresser.decompressPlugin=decompressPlugin;
-            decompresser.decompressHandle=decompressPlugin->dictDecompressOpen(decompressPlugin,1,uncompressDataSize,0,0);
+            decompresser.decompressHandle=decompressPlugin->dictDecompressOpen(decompressPlugin,1,(size_t)uncompressDataSize,0,0);
             check((decompresser.decompressHandle!=0),kSyncClient_decompressOpenError);
             check(decompressPlugin->dictDecompress(decompresser.decompressHandle,0,codeBuf,codeBuf+compressDataSize,
                                                    decompressBuf,decompressBuf+uncompressDataSize),
